requirements.txt
opencv-python
opencv-contrib-python
torch
numpy
poetry
# dlib
ultralytics
natsort
streamlit
pyaudio
SpeechRecognition
pymongo
pandas
python-dotenv